src/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv>=1.0.0
mlflow>=2.18.0
openai>=1.0.0
litellm>=1.40.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
prometheus-client==0.20.0
python-multipart>=0.0.6
qdrant-client>=1.7.0
httpx>=0.25.0
